Rearranged Composer components
