+++ src/app/components/Composer/constants.js
@@ -1,0 +1,55 @@
+// tools
+import React from "react"
+import { Raw } from "slate"
+
+// plugins
+import { MarkHotkey } from "./plugins"
+
+
+// composer initial state
+export const headerPlaceholders = {
+	"titlePlaceholder": "Write Your Title Here ✍",
+	"subtitlePlaceholder": "Subtitle (Optional)"
+}
+export const initialState = Raw.deserialize({
+  nodes: [
+    {
+      kind: "block",
+      type: "paragraph",
+      nodes: [
+        {
+          kind: "text",
+          text: "🌠 You can drag-and-drop your image(s) here. Please add a caption to each (title, date, place, camera etc.) "
+        },
+        {
+          kind: "text",
+          text: "<strong>Make this a photo essay, image series, article or just a single photo submission.</strong>",
+          //mark: "bold"
+        }
+      ]
+    }
+  ]
+}, { terse: true })
+
+// state schema
+export const stateSchema = {
+	nodes: {
+		code: props => <pre {...props.attributes}><code>{props.children}</code></pre>
+	},
+	marks: {
+		bold: props => <strong>{props.children}</strong>,
+		code: props => <code>{props.children}</code>,
+		italic: props => <em>{props.children}</em>,
+		strikethrough: props => <del>{props.children}</del>,
+		underline: props => <u>{props.children}</u>,
+	}
+}
+
+// plugins list
+export const plugins = [
+  MarkHotkey({ key: 'b', type: 'bold' }),
+  MarkHotkey({ key: 'c', type: 'code', isAltKey: true }),
+  MarkHotkey({ key: 'i', type: 'italic' }),
+  MarkHotkey({ key: 'd', type: 'strikethrough' }),
+  MarkHotkey({ key: 'u', type: 'underline' })
+]
