+++ src/app/components/Composer/plugins/index.js
