--- src/app/views/Submit/Composer/styles.js
@@ -1,17 +1,0 @@
-// tools
-import styled from "styled-components"
-
-// css
-export const Wrapper = styled.div`
-// 	.public-DraftEditorPlaceholder-root {
-// 		position: 	absolute;
-// 		z-index: 		${ props => props.theme.layer.tuck };
-// 		opacity: 		${ props => props.theme.opacity.half };
-// 	}
-// 	.DraftEditorPlaceholder-hidden { display: none; }
-// 	.public-DraftEditor-content {
-// 		height: 							100vh;
-// 		-webkit-user-modify: 	read-write-plaintext-only; // remove iOS edit buttons
-// 		user-modify: 					read-write-plaintext-only; // remove iOS edit buttons
-// 	}
-`
